Adjusted social icons, added 404 error page, adjusted app bar
