+++ package.json
@@ -45,7 +45,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-redux": "^8.0.5",
-    "react-router-dom": "^6.10.0",
+    "react-router-dom": "^6.11.0",
     "sass": "^1.60.0",
     "styled-components": "^5.3.9",
     "webpack": "^5.77.0"
