+++ public/index.html
@@ -7,6 +7,9 @@
     <meta name="Grandma's Cookbook" content="" />
     <link rel="stylesheet" type="text/css" href="../client/index.css">
     <script defer="defer" src="/bundle.js"></script>
+    <style>
+      @import url("https://fontsforyou.com/fonts/h/Herculanum.ttf");
+  </style>
     <title>Grandma's Cookbook</title>
 
     <!-- To ensure webpack-dev-server work with react-router, all <script>/<link> tags
